html/browsers/browsing-the-web/navigating-across-documents/012-timeout.html

<!doctype html>
<!--/
specs:
  - id: browsing-the-web.html#navigating-across-documents
    steps: [6]

description: |
  1. This test creates an anchor with both an onclick and href.
  2. The test "clicks" the anchor, triggering the onclick handler operation,
      which will attempt to document.write (after a 2s delay), followed by
      navigating the iframe to the anchor's `href=""`.

/-->
<title>Link with onclickwith delayed (timeout) document.write and href navigation</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<iframe id="test" name="test"></iframe>
<a target="test" onclick="setTimeout(() => document.write('<script>parent.postMessage(&quot;write&quot;, &quot;*&quot;)</script>'), 2000);" href="href.html">Test</a>
<script>

var t = async_test(undefined, {
  timeout: 4000
});

t.step(function() {
  document.getElementsByTagName("a")[0].click();
});

onmessage = t.step_func(function(e) {
  assert_equals(e.data, "href");
  t.done();
});

</script>
